배포 준비 및 문구 변경

+++ client/src/utils/crypto.ts
@@ -150,7 +150,7 @@ export async function deriveKeyFromPassword(
     
     return {
         key,
-        salt: btoa(String.fromCharCode(...salt)),
+        salt: btoa(String.fromCharCode(...(salt as Uint8Array))),
     };
 }
